extract layout and UI primitives from app.js to core/ui.js
