+++ public/app.js
@@ -1,6 +1,7 @@
 // Extracted helpers
 import { formatTitle, $, h, fmtSize, fmtDate, formatTimestamp } from '/core/helpers.js';
 import { api, channelCover, videoThumb, videoUrl } from '/core/api.js';
+import { renderLayout, pagination, lazyThumbs } from '/core/ui.js';
 // --- tiny router (hash-based) ---
 const routes = {
   "": renderHome,
@@ -411,12 +412,7 @@ async function toggleFav(e, item) {
 // API + URL helpers moved to /core/api.js
 
 // --- layout & header ---
-function renderLayout(content){
-  const app = $("#app");
-  app.innerHTML = "";
-  const container = h("div", { class: "container" }, content);
-  app.append(container);
-}
+// layout moved to /core/ui.js
 
 // --- Theme logic ---
 async function getThemeSetting() {
@@ -1071,37 +1067,10 @@ function rowVideo(channelName, v) {
 //   );
 // }
 
-function pagination(meta, onPage){
-  const btn = (label, p, disabled=false) =>
-    h("button", { class: "pagination-btn", disabled, onclick: ()=> onPage(p) }, label);
-
-  return h("div", { class: "pagination" },
-    btn("« First", 1, meta.page === 1),
-    btn("‹ Prev", Math.max(1, meta.page-1), meta.page === 1),
-    h("span", { class: "cur" }, `Page ${meta.page} / ${meta.totalPages}`),
-    btn("Next ›", Math.min(meta.totalPages, meta.page+1), meta.page === meta.totalPages),
-    btn("Last »", meta.totalPages, meta.page === meta.totalPages)
-  );
-}
+// pagination moved to /core/ui.js
 
 // --- lazy thumbnails ---
-function lazyThumbs(){
-  const imgs = document.querySelectorAll("img.lazy");
-  const io = new IntersectionObserver(entries=>{
-    for (const ent of entries) {
-      if (ent.isIntersecting) {
-        const img = ent.target;
-        const src = img.getAttribute("data-src");
-        if (src) {
-          img.src = src;
-          img.removeAttribute("data-src");
-          io.unobserve(img);
-        }
-      }
-    }
-  }, { rootMargin: "300px" });
-  imgs.forEach(i => io.observe(i));
-}
+// lazyThumbs moved to /core/ui.js
 
 // --- router hook ---
 function onRoute(){
